Module rowantlr::backend::dfa

DFA related data structures, backend for regular expressions.

Use Expr::build, BuildResult::try_resolve to build DFAs;

Run a DFA on some input with Dfa::run and Resolved::run;

Minimise a DFA with Resolved::minimise;

Below is an example:

use rowantlr::ir::lexical::Expr;
use rowantlr::backend::dfa::InvalidInput;
let expr = Expr::concat([
    Expr::singleton('a'),
    Expr::many(Expr::union([
        Expr::singleton('a'),
        Expr::singleton('b'),
    ])),
    Expr::singleton('b'),
]);
let resolved = expr.build().try_resolve().unwrap();
// the minimised DFA and the original should behave the same
for m in [resolved.clone(), resolved.minimise()] {
    assert_eq!(Some(()), m.run("aab".chars()).unwrap());
    assert_eq!(Some(()), m.run("ab".chars()).unwrap());
    assert_eq!(Some(()), m.run("abab".chars()).unwrap());
    assert_eq!(None, m.run("aa".chars()).unwrap());
    let InvalidInput { current_state, current_input, remaining_input }
        = m.run("baaa".chars()).unwrap_err();
    assert_eq!(current_state, 0);
    assert_eq!(current_input, 'b');
    assert_eq!(remaining_input.as_str(), "aaa");
}

Use Expr::build_many to build Dfa from multiple regular expressions, and use BuildResult::apply_hint to resolve a conflict (BuildResult::apply_hints can also be used to resolve conflicts in batch):

use rowantlr::ir::lexical::{Expr, PosInfo};
use rowantlr::backend::dfa::ResolveError;
// start with 'a', end with 'b'
let e1 = Expr::concat([
    Expr::singleton('a'),
    Expr::many(Expr::any_of("ab")),
    Expr::singleton('b'),
]);
// even number of 'a's and 'b's.
let e2 = Expr::many(Expr::union([
    Expr::from("aa"),
    Expr::from("bb"),
    Expr::concat([
        Expr::from("ab") | Expr::from("ba"),
        Expr::many(Expr::from("aa") | Expr::from("bb")),
        Expr::from("ab") | Expr::from("ba"),
    ]),
]));
// 'e1' and 'e2' should intersect.
let result = Expr::build_many([(&e1, 0), (&e2, 1)]);
let (result, conflicts) = result.try_resolve().expect_err("conflict expected");
// exactly one conflict is detected.
assert_eq!(conflicts.len(), 1);
let conflict = &conflicts[0];
// tag '0' and tag '1' are involved in this conflict.
assert_eq!(vec![0, 1], conflict.conflicting_tags(&result).copied().collect::<Vec<_>>());
// as for the different interpretations for some problematic input ...
conflict.interpretations.iter().for_each(|ps| {
    // all interpretations is indeed of the same input "aabb".
    assert_eq!("aabb".to_string(), ps.split_last().unwrap().1.iter()
        .map(|&p| result.position_info[p as usize].info.into_normal().unwrap())
        .collect::<String>());
    // each position list is chained according to the 'follow_pos' relation.
    ps.windows(2).for_each(|p|
        assert!(result.position_info[p[0] as usize].follow_pos.contains(&p[1])));
});
// apply a hint to resolve the conflict.
let mut result = result;
match result.apply_hint("aabb".chars(), 0) {
    ResolveError::Resolved(s) => {
        // we indeed resolved the conflict between tag '0' and tag '1' for input "aabb".
        assert_eq!(2, s.len());
        assert!(matches!(result.position_info[s[0] as usize].info, PosInfo::Accept(0)));
        assert!(matches!(result.position_info[s[1] as usize].info, PosInfo::Accept(1)));
    }
    err => panic!(r#""aabb" should resolve a conflict, but we got {:?} instead"#, err),
}
// now we are ready to build the DFA.
let resolved = result.try_resolve().expect("no more conflict expected");
// the minimised DFA and the original should behave the same
for m in [resolved.clone(), resolved.minimise()] {
    // the following two inputs have no conflict from the very beginning:
    assert_eq!(Some(0), m.run("abb".chars()).unwrap());
    assert_eq!(Some(1), m.run("bbaa".chars()).unwrap());
    // the hint we used to resolve the conflict behaves well:
    assert_eq!(Some(0), m.run("aabb".chars()).unwrap());
    // other conflicts are solved just as for the hint:
    assert_eq!(Some(0), m.run("abab".chars()).unwrap());
}

The above two example DFAs are actually minimal from the very beginning. Below is an example of how the DFA can have equivalent states and be properly minimised:

use rowantlr::ir::lexical::Expr;
let expr = Expr::union([
    Expr::some(Expr::singleton('a')), // a+
    Expr::some(Expr::from("aa")),     // (aa)+
]);
let resolved = expr.build().try_resolve().unwrap();
let minimised = resolved.clone().minimise();
// the minimised DFA and the original should behave the same
for m in [&resolved, &minimised] {
    assert_eq!(Some(()), m.run("a".chars()).unwrap());
    assert_eq!(Some(()), m.run("aa".chars()).unwrap());
    assert_eq!(Some(()), m.run("aaa".chars()).unwrap());
    assert_eq!(None, m.run("".chars()).unwrap());
}
assert!(minimised.dfa.state_count < resolved.dfa.state_count);
assert_eq!(resolved.dfa.state_count, 3);
assert_eq!(minimised.dfa.state_count, 2);

Structs

BuildResult

Result from building a Dfa.

Builder

Calculate followpos(p) for every position p in a regular expression e. This piece of information can be used to construct a DFA.

Conflict

Conflict in DFA states.

Dfa

Definite Finite State Automata.

InvalidInput

Invalid input for some Dfa.

NotAcceptState

Decide what to do next for a non-accepting state.

PosEntry

Information related to a position in a regular expression.

Reachability

Information for reachability for states in a DFA for a specific BuildResult.

Resolved

BuildResult with conflicts resolved.

Enums

ResolveError

Result for resolving conflicts.