Module rowantlr::backend::lalr1

LALR(1) backend for rowantlr and related stuffs.

Prerequisites (as illustrated in module ll1):

use the Grammar API to create new grammars.

use calc_deduce_to_empty to calculate DEDUCE_TO_EMPTY.

use calc_first to calculate FIRST.

Call build to generate ACTION and GOTO tables, and use Table::simulate_parse to test the grammar on some token streams.

#![allow(non_snake_case)]
use std::convert::TryFrom;
use rowantlr::r#box;
use rowantlr::ir::syntax::{Grammar, epsilon, Symbol::*, Expr};
use rowantlr::backend::{ll1, ll1::Lookahead, lalr1, lalr1::Action::*};

let mut g = Grammar::<&'static str>::build(|g| {
    let [S, L, R] = g.add_non_terminals();
    // S —→ L = R | R
    g.mark_as_start(S);
    g.add_rule(S, r#box![NonTerminal(L), Terminal("="), NonTerminal(R)]);
    g.add_rule(S, r#box![NonTerminal(R)]);
    // L —→ * R | id
    g.add_rule(L, r#box![Terminal("*"), NonTerminal(R)]);
    g.add_rule(L, r#box![Terminal("id")]);
    // R —→ L
    g.add_rule(R, r#box![NonTerminal(L)]);
});

let deduce_to_empty = ll1::calc_deduce_to_empty(&g);
let first = ll1::calc_first(&g, &deduce_to_empty);
let kernels = lalr1::build(&g, &first, &deduce_to_empty);
let table = lalr1::Table::try_from(&kernels).ok().unwrap();
// correctly parse "* * id = id"
let parse = table.simulate_parse(["*", "*", "id", "=", "id"]).unwrap();
assert_eq!(
    format!("{}", parse.pretty(&["S'", "S", "L", "R"])),
    indoc::indoc! {r#"
        NT#[ S ]
          NT#[ L ]
            "*"
            NT#[ R ]
              NT#[ L ]
                "*"
                NT#[ R ]
                  NT#[ L ]
                    "id"
          "="
          NT#[ R ]
            NT#[ L ]
              "id"
    "#});
// cannot parse "* * id id" (expecting "=" or EOF)
let err = table.simulate_parse(["*", "*", "id", "id"]).unwrap_err();
assert_eq!(err.unexpected, Lookahead::new("id"));
assert_eq!(err.expecting, r#box![Lookahead::END_OF_INPUT, Lookahead::new("=")]);
// cannot parse "* * id =" (expecting "id" or "*")
let err = table.simulate_parse(["*", "*", "id", "="]).unwrap_err();
assert_eq!(err.unexpected, Lookahead::END_OF_INPUT);
assert_eq!(err.expecting, r#box![Lookahead::new("*"), Lookahead::new("id")]);

To visualise the generated LALR automaton, run the following:

println!("{}", kernels.display_dot2tex(&["S'", "S", "L", "R"]));

and typeset the generated contents using doc2tex and LaTeX.

Modules

simple

So-said “simple” types, with the tag types muted.

simulation

Simulation of LALR(1) parsing.

Structs

Entry

An LALR entry in a state set.

FrozenKernel

Kernel sets frozen for efficient access.

KernelSets

All kernel sets for a Grammar, together with the full GOTO table.

LalrConflict

Conflicts in LALR kernel sets.

Lookaheads

A set of Lookahead tokens.

RawGotoSets

Iterables for traversing GOTO sets, always call RawGotoSets::get on this type.

Rule

A production rule used in a state set.

Table

The LALR parsing table.

Enums

Action

An entry of the ACTION table.

Traits

TagManager

Tag managers control the behaviour of closure when an Entry is to be generated.

TagResolver

Tags in LALR state sets can be resolved to eliminate their inter-dependency.

Functions

all_goto_sets

Calculate all non-empty GOTO(I, X) for each X ∈ V.

all_kernel_sets

Calculate all kernel sets of a given Grammar.

build

Build LALR kernel sets table, and the GOTO table.

closure

The CLOSURE of an item set.

resolve_tags

Resolve the tags with a TagResolver in case their is inter-dependencies among them.

Type Definitions

Kernel

An LALR kernel state set (all A -> . β rules omitted). This property is not enforced, this type is only for better readability.

NonTerminal

Non-terminal symbols, represented as their index in the Grammar, including the internal start symbol S (indexed 0).

State

An LALR state set.