Trait rowantlr::backend::lalr1::TagManager

pub trait TagManager<A> {
    type Tag: 'static;
    fn generate_tag(
        &mut self, 
        upcoming: &[Symbol<A>], 
        parent_tag: &Self::Tag
    ) -> Self::Tag;

    fn update_tag(
        &mut self, 
        upcoming: &[Symbol<A>], 
        parent_tag: &Self::Tag, 
        previous: &Self::Tag
    );

    fn merge_tags<'a>(
        &mut self, 
        tags: impl Iterator<Item = (&'a Self::Tag, &'a Self::Tag)>
    );

    fn root_tag(&mut self) -> Self::Tag;
}

Tag managers control the behaviour of closure when an Entry is to be generated.

The same non-terminal symbol might be generated repeatedly from different production rules, and when this happens, we need to somehow update the previously-generated Tag. However, the way we store the tags makes it impossible to get mutable references to a Tag, so the method TagManager::update_tag is only provided a shared reference.

There are two possible strategies for dealing with such situations:

Keep all the real contents of the tags inside this TagManager, and use handles (possibly indices) into this manager for tag types. This way when a tag is to be updated, just modify the real contents in this manager, and leave the tags (handles) untouched.

Wrap tag types with Cells (when Tag is Copy) or RefCells (when Tags are heavy) etc. to obtain inherent mutability. This way two Tags can be properly updated.

Associated Types

type Tag: 'static

The tag type the Entrys is about to use.

Required methods

fn generate_tag(
&mut self,
upcoming: &[Symbol<A>],
parent_tag: &Self::Tag
) -> Self::Tag

For the first time we see an entry shaped A -> α . X β (where α, β ∈ V* and X ∈ VN), this method is used to generate a new Tag for this entry. The tag for A is passed in as the parent_tag, and β is provided as upcoming symbols.

fn update_tag(
&mut self,
upcoming: &[Symbol<A>],
parent_tag: &Self::Tag,
previous: &Self::Tag
)

If a Tag is already present for an entry, and this entry is generated again possibly by some different production rule in the set, this method is used to update the old Tag.

fn merge_tags<'a>(
&mut self,
tags: impl Iterator<Item = (&'a Self::Tag, &'a Self::Tag)>
)

After generating the tags for a state set, we might discover that the set is duplicated. In this case this method is used to merge the tags from the two sets, and feel free to remove those generated tags.

The tags are presented as a series of pairs.

fn root_tag(&mut self) -> Self::Tag

The entries generated from the production rule of the start symbol S do not have a valid parent, nor do they have upcoming symbols. Tags for them are generated specially here.