Module rowantlr::backend::ll1

LL(1) backend for rowantlr, and some common utilities.

For each non-terminal N:

DEDUCE_TO_EMPTY(N) indicates whether N —→* ε.

FIRST(N) collects all terminal a such that N —→* a β (β ∈ V*).

Below is an illustration, one can:

use the Grammar API to create new grammars.

use calc_deduce_to_empty to calculate DEDUCE_TO_EMPTY.

use calc_first to calculate FIRST.

#![allow(non_snake_case)]
use rowantlr::r#box;
use rowantlr::ir::syntax::{Grammar, epsilon, Symbol::*};
use rowantlr::backend::ll1::{calc_deduce_to_empty, calc_first};

let mut g = Grammar::<&'static str>::build(|g| {
    let [E, T, E_, T_, F] = g.add_non_terminals();
    // E  —→ T E'
    g.mark_as_start(E);
    g.add_rule(E, r#box![NonTerminal(T), NonTerminal(E_)]);
    // E' —→ + T E' | ε
    g.add_rule(E_, r#box![Terminal("+"), NonTerminal(T), NonTerminal(E_)]);
    g.add_rule(E_, epsilon());
    // T  —→ F T'
    g.add_rule(T, r#box![NonTerminal(F), NonTerminal(T_)]);
    // T' —→ * F T' | ε
    g.add_rule(T_, r#box![Terminal("*"), NonTerminal(F), NonTerminal(T_)]);
    g.add_rule(T_, epsilon());
    // F  —→ ( E ) | id
    g.add_rule(F, r#box![Terminal("("), NonTerminal(E), Terminal(")")]);
    g.add_rule(F, r#box![Terminal("id")]);
});

// Calculate `DEDUCE_TO_EMPTY` and `FIRST`.
let deduce_to_empty = calc_deduce_to_empty(&g);
assert_eq!(deduce_to_empty, r#box![false, false, false, true, true, false]);
let first = calc_first(&g, &deduce_to_empty);
assert_eq!(first, r#box![
    r#box!["(", "id"],
    r#box!["(", "id"],
    r#box!["(", "id"],
    r#box!["+"],
    r#box!["*"],
    r#box!["(", "id"],
]);

Structs

Lookahead

A lookahead token.

Functions

append_first_of

Append FIRST(β) to a given BTreeSet for any string β ∈ V*, according to the provided FIRST and DEDUCE_TO_EMPTY sets. The returned bool indicates whether or not this input string may deduce to empty.

calc_deduce_to_empty

Calculate the DEDUCE_TO_EMPTY set for each non-terminal.

calc_first

Calculate the FIRST set for each non-terminal.

first_of

Calculate FIRST(β) for any string β ∈ V*, according to the provided FIRST and DEDUCE_TO_EMPTY sets. The bool indicates whether or not this string may deduce to empty.