Struct rowantlr::ir::lexical::Expr

pub struct Expr<A>(_);

Regular expressions.

Implementations

impl Expr<CharClass>

pub fn freeze_char_class<'a, I>(exprs: I) -> (Vec<Expr<u32>>, Dict<(Char, u32)>) where
I: IntoIterator<Item = &'a Expr<CharClass>>,
I::IntoIter: Clone,

Analyse the CharClasses, and produce a character classifier and an Expr with char classes replaced by their indices in the classifier.

use rowantlr::ir::lexical::char_class::Char;
// auxiliary function for constructing a singleton of a char class:
fn char_class<C: Into<CharClass>>(c: C) -> Expr<CharClass> { Expr::singleton(c.into()) }
// a fictitious lexical grammar for identifiers
let mod_id_start = Expr::union([
    char_class('A'..='Z'),
    char_class('$'),
]);
let var_id_start = Expr::union([
    char_class('a'..='z'),
    char_class('_'),
]);
let id_cont = Expr::many(Expr::union([
    char_class('a'..='z'),
    char_class('A'..='Z'),
    char_class('_'),
    char_class('0'..='9'),
    char_class('\''),
]));
let mod_id = Expr::concat([mod_id_start, id_cont.clone()]);
let var_id = Expr::concat([var_id_start, id_cont]);
let (ids, inputs) = Expr::freeze_char_class([&mod_id, &var_id]);
// the expressions with input frozen
assert_equal(ids, [
    Expr::concat([
        Expr::union_of([1, 3]),
        Expr::many(Expr::union_of([2, 3, 4])),
    ]),
    Expr::concat([
        Expr::singleton(4),
        Expr::many(Expr::union_of([2, 3, 4]))
    ]),
]);
// auxiliary function for finding the next char:
fn next_char(c: char) -> char { char::from_u32(c as u32 + 1).unwrap() }
// the input classifier
assert_eq!(inputs, [
    ('\0', 0),
    ('$',  1), (next_char('$'),  0),
    ('\'', 2), (next_char('\''), 0),
    ('0',  2), (next_char('9'),  0),
    ('A',  3), (next_char('Z'),  0),
    ('_',  4), (next_char('_'),  0),
    ('a',  4), (next_char('z'),  0),
].into_iter().map(|(c, k)| (Char::from(c), k)).collect());

For usage of the resulting classifier, see Dict::classify.

impl<A> Expr<A>

pub const fn singleton(a: A) -> Expr<A>

Singleton: a.

pub fn epsilon() -> Expr<A>

Empty string: ε.

pub fn union(xs: impl IntoIterator<Item = Expr<A>>) -> Expr<A>

Union: x₁ | x₂ | ... | xₙ.

If input has only one branch, no extra Op::Union layer is introduced:

let e = Expr::from("abc");
assert_eq!(e.clone(), Expr::union([e]));

Nested Op::Union structures will be simplified:

assert_eq!(Expr::union([
    Expr::singleton('a'),
    Expr::any_of("bc"),
    Expr::singleton('d'),
]), Expr::any_of("abcd"));

Note that Expr::any_of is just a convenient method for Expr<char>, a thin wrapper around Expr::union and Expr::singleton. Here is the fully expanded version:

assert_eq!(Expr::union([
    Expr::singleton('a'),
    Expr::union([
        Expr::singleton('b'),
        Expr::singleton('c'),
    ]),
    Expr::singleton('d'),
]), Expr::union([
    Expr::singleton('a'),
    Expr::singleton('b'),
    Expr::singleton('c'),
    Expr::singleton('d'),
]));

pub fn concat(xs: impl IntoIterator<Item = Expr<A>>) -> Expr<A>

Concatenation: x₁ x₂ ... xₙ.

If input has only one item, no extra Op::Concat layer is introduced:

let e = Expr::singleton('a');
assert_eq!(e.clone(), Expr::concat([e]));

Nested Op::Concat structures will be simplified:

assert_eq!(Expr::concat([
    Expr::singleton('a'),
    Expr::from("bc"),
    Expr::singleton('d'),
]), Expr::from("abcd"));

Note that Expr::from is just a convenient method for Expr<char>, a thin wrapper around Expr::concat and Expr::singleton. Here is the fully expanded version:

assert_eq!(Expr::concat([
    Expr::singleton('a'),
    Expr::concat([
        Expr::singleton('b'),
        Expr::singleton('c'),
    ]),
    Expr::singleton('d'),
]), Expr::concat([
    Expr::singleton('a'),
    Expr::singleton('b'),
    Expr::singleton('c'),
    Expr::singleton('d'),
]));

pub fn some(expr: Expr<A>) -> Expr<A>

Positive closure: x+.

pub fn optional(expr: Expr<A>) -> Expr<A>

Optional: x? = ε | x.

pub fn many(expr: Expr<A>) -> Expr<A>

Closure, or the Kleene star: x* = ε | x+.

impl Expr<char>

pub fn any_of(s: &str) -> Self

Union of all the characters in some string.

impl<A> Expr<A>

pub fn union_of<I: IntoIterator<Item = A>>(s: I) -> Self

Union of all the elements in some iterator.

impl<A: Clone> Expr<A>

pub fn traverse<V: ExprVisitor<A>>(&self, visitor: &mut V) -> V::Output

Traverse the expression tree, building up ExprInfo on the fly, and use the information during the traversal.

pub fn traverse_extended<Tag, V: ExprVisitorExt<A, Tag>>(
&self,
visitor: &mut V,
tag: Tag
) -> V::Output

For regular expression e, collect the information as if we are traversing the extended regular expression e# (where # ∉ Σ).

impl<A: Ord + Clone> Expr<A>

pub fn build(&self) -> BuildResult<A, ()>

Build a Dfa from a regular expression. See also Builder::build.

pub fn build_many<'a, I, Tag>(exprs: I) -> BuildResult<A, Tag> where
I: IntoIterator<Item = (&'a Expr<A>, Tag)>,
A: 'a,

Build a Dfa from many regular expressions. See also Builder::build.