Module rowantlr::utils::partition_refinement

The partition refinement algorithm.

The algorithm is implemented by maintaining the following information:

all the sets, associated with each set S its collection of elements.

all the elements, associated with each element the set it belongs to.

These requirements are captured in structure Partitions and trait IndexManager. For most cases, the n elements are representable by 0..n, and thus TrivialIdxMan can be used.

Example

let mut partitions = Partitions::<u64>::new(vec![0, 1, 2, 3, 4, 5, 6]);
partitions.refine_with([1, 3, 5], &mut ()); {
    assert_eq!(partitions.parts().len(), 2);
    assert_equal(partitions.parts(), [&[0, 6, 2, 4], &[5, 3, 1]]);
}
partitions.refine_with([4, 5, 6], &mut ()); {
    assert_eq!(partitions.parts().len(), 4);
    assert_equal(partitions.parts(), [&[0, 2], &[1, 3], &[6, 4], &[5]]);
}
partitions.refine_with([0, 2, 4], &mut ()); {
    assert_eq!(partitions.parts().len(), 5);
    assert_equal(partitions.parts(), [&[2, 0], &[1, 3], &[6], &[5], &[4]]);
}
// random access to parts is also possible
assert_eq!(partitions.part(1), &[1, 3]);
assert_eq!(partitions.part(4), &[4]);

Use Intervals to collect all newly-generated parts during the refinement:

let mut partitions = Partitions::<u64>::new_trivial(7);
let mut intervals = Intervals::new();
partitions.refine_with([1, 3, 5], &mut intervals); { // 0 6 2 4 | 5 3 1
    assert_equal(&intervals, &[4..7]);
    assert_equal(intervals.parts(&partitions).into_elements(), [&[5, 3, 1]]);
}
partitions.refine_with([4, 5, 6], &mut intervals); { // 0 2 | 6 4 | 1 3 | 5
    assert_equal(&intervals, &[0..2, 4..7]);
    assert_equal(intervals.parts(&partitions).into_elements(),
                 [&[1_u64, 3] as &[u64], &[5], &[0, 2]]);
}
partitions.refine_with([0, 2, 4], &mut intervals); { // 2 0 | 6 | 4 | 1 3 | 5
    assert_equal(&intervals, &[0..3, 4..7]);
    assert_equal(intervals.parts(&partitions).into_elements(),
                 [&[1_u64, 3] as &[u64], &[5], &[2, 0], &[6]]);
}

Note that the intervals are only meaningful when interpreted together with the partitions. These intervals are only exposed for debug purposes.

Structs

Part

A sub set of elements in partition refinement. Never invalidated after generation.

Partitions

Perform partition refinement on some elements.

TrivialIdxMan

A trivial IndexManager for bounded integral elements (essentially 0..n).

Traits

IndexManager

Maintain the reverse mapping from elements in a slice to their indices.

PartManager

Part managers controls the behaviour of Partitions::refine_with when new parts are generated during the refinement. For a “do-nothing” manager, use ().