Trait rowantlr::utils::tuple::TupleBorrow

pub trait TupleBorrow<'a> {
    type Borrowed;
    fn tuple_borrow(&'a self) -> Self::Borrowed;
}

Get a shared reference for each element of a tuple.

Examples

let data: (String, Vec<&'static str>, u64) =
    ("answer".to_string(), vec!["life", "the universe", "everything"], 42);
let borrowed: (&String, &Vec<&'static str>, &u64) = data.tuple_borrow();

Associated Types

type Borrowed

A tuple type shaped as Self, with each element replaced with its shared reference.

Required methods

fn tuple_borrow(&'a self) -> Self::Borrowed

Borrow each element of self.